Refactor JsonXExtensionsSpecBuilder with pulling each builder into each own class
